David/Pat Move translation creation logic into TxcMakerParser
