+++ src/main/java/org/mercycorps/translationcards/txcmaker/model/NewDeck.java
@@ -1,9 +1,6 @@
 package org.mercycorps.translationcards.txcmaker.model;
 
-import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 
 public class NewDeck {
     private String sourceLanguage;
@@ -16,7 +13,15 @@ public class NewDeck {
     private List<Translation> translations;
     private List<String> destinationLanguageNames;
 
-    public NewDeck(String sourceLanguage, String deckLabel, String author, long timestamp, boolean locked, String license_url, String readme, List<NewCard> cards) {
+    public NewDeck(String sourceLanguage,
+                   String deckLabel,
+                   String author,
+                   long timestamp,
+                   boolean locked,
+                   String license_url,
+                   String readme,
+                   List<Translation> translations,
+                   List<String> destinationLanguageNames) {
         this.sourceLanguage = sourceLanguage;
         this.deckLabel = deckLabel;
         this.author = author;
@@ -24,34 +29,7 @@ public class NewDeck {
         this.locked = locked;
         this.license_url = license_url;
         this.readme = readme;
-        this.translations = buildTranslationsFromCards(cards);
-        this.destinationLanguageNames = buildDestinationLanguageNames(cards);
-    }
-
-    List<String> buildDestinationLanguageNames(List<NewCard> cards) {
-        Map<String, Object> destinationLanguageNames = new HashMap<>();
-        for (NewCard card : cards) {
-            destinationLanguageNames.put(card.getDestinationLanguageName(), null);
-        }
-        return new ArrayList<>(destinationLanguageNames.keySet());
-    }
-
-    List<Translation> buildTranslationsFromCards(List<NewCard> cards) {
-        Map<String, Translation> sourcePhraseToTranslation = new HashMap<>();
-        for (NewCard card : cards) {
-            String sourcePhrase = card.getSourcePhrase();
-            if (sourcePhraseToTranslation.containsKey(sourcePhrase)) {
-                sourcePhraseToTranslation.get(sourcePhrase).addCard(card);
-            } else {
-                sourcePhraseToTranslation.put(sourcePhrase, createTranslationFromCard(card));
-            }
-        }
-        return new ArrayList<>(sourcePhraseToTranslation.values());
-    }
-
-    private Translation createTranslationFromCard(NewCard card) {
-        List<NewCard> cardListForTranslation = new ArrayList<>();
-        cardListForTranslation.add(card);
-        return new Translation(cardListForTranslation);
+        this.translations = translations;
+        this.destinationLanguageNames = destinationLanguageNames;
     }
 }
